refactor(scroll): one register_scroll_nodes, not three
There were three answers to "which nodes are scrollable and how big are
they": the dll's `register_scroll_nodes`, a hand-maintained PORT of it in the
e2e runner, and whatever each `layout/tests` fixture hand-seeded through
`update_node_bounds`.

That is the divergence that let scroll bugs hide from the suite. A test could
park a scroll state production would never produce — and did: the caret-reveal
tests seeded container and content rects of their own choosing, so they
exercised the reveal against geometry the engine never computes. A fix in one
host could also be left standing in the other, which is exactly what happened
to the VirtualView necessity amendment (both copies carry the same comment
explaining why it must be there).

The body moves to `azul_layout::managers::scroll_registration`, taking `now`
as a parameter rather than reading the wall clock, so a frozen-clock test can
drive it. All three call it: the dll, the e2e runner, and tests.

`caret_follows_typing.rs` is switched over as the first consumer — it now
publishes its scroll containers exactly the way production does, so its
fixture cannot drift from the engine. Both tests still pass, which is the
point: the reveal was being measured against real geometry all along, and now
that is enforced rather than coincidental.

azul-layout --lib 7282 pass; azul-dll --lib 1907 pass.
